% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgan.R
\name{ctgan_sample}
\alias{ctgan_sample}
\title{Synthesize Data Using a CTGAN Model}
\usage{
ctgan_sample(ctgan_model, n = 100)
}
\arguments{
\item{ctgan_model}{A fitted `CTGANModel` object.}

\item{n}{Number of rows to generate.}
}
\description{
Synthesize Data Using a CTGAN Model
}
