% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-load.R
\name{ctgan_save}
\alias{ctgan_save}
\title{Save a CTGAN model}
\usage{
ctgan_save(ctgan_model, path)
}
\arguments{
\item{ctgan_model}{A fitted `CTGANModel` object.}

\item{path}{Path to which the save the model.}
}
\description{
Serialize a CTGAN Model object to disk.
}
