% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgan.R
\name{fit.CTGANModel}
\alias{fit.CTGANModel}
\title{Train a CTGAN Model}
\usage{
\method{fit}{CTGANModel}(object, train_data, epochs = 100, log_frequency = TRUE, ...)
}
\arguments{
\item{object}{A `CTGANModel` object.}

\item{train_data}{Training data, should be a data frame.}

\item{epochs}{Number of epochs to train.}

\item{log_frequency}{Whether to use log frequency of categorical levels in
conditional sampling. Defaults to `TRUE`.}

\item{...}{Additional arguments, currently unused.}
}
\description{
Train a CTGAN Model
}
