% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_ctgan}
\alias{install_ctgan}
\title{Install CTGAN}
\usage{
install_ctgan(
  method = c("auto", "virtualenv", "conda"),
  conda = "auto",
  envname = NULL,
  restart_session = TRUE,
  ...
)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds
a method that will work in the local environment.}

\item{conda}{Path to conda executable (or "auto" to find conda using the
\code{PATH} and other conventional install locations).}

\item{envname}{Name of Python environment to install within.}

\item{restart_session}{Whether to restart R session after installation.}

\item{...}{Additional parameters.}
}
\description{
Installs the CTGAN python package.
}
\details{
This model also requires Python 3, Python 2 is unsupported.
}
