% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgan.R
\name{ctgan}
\alias{ctgan}
\title{Initialize a CTGAN Model}
\usage{
ctgan(
  embedding_dim = 128,
  gen_dim = c(256, 256),
  dis_dim = c(256, 256),
  l2_scale = 1e-06,
  batch_size = 500
)
}
\arguments{
\item{embedding_dim}{Dimension of embedding layer.}

\item{gen_dim}{Dimensions of generator layers.}

\item{dis_dim}{Dimensions of discriminator layers.}

\item{l2_scale}{ADAM weight decay.}

\item{batch_size}{Batch size.}
}
\description{
Initializes a CTGAN model object
}
